+++ src/com/elka/nn/NeuralNet.java
@@ -15,16 +15,16 @@ public class NeuralNet {
 	 * private final int NODES_NUM = 6; // wektor 5 danych + 1 polaryzacja
 	 * private final int SINX_NODES = 2; // jedna dana + 1 polaryzacja
 	 */
-	private float error;
-	private float prevError;
-	private float learnRate;
+	private double error;
+	private double prevError;
+	private double learnRate;
 
-	private final float KW = 1.01f;
-	private final float PD = 0.7f;
-	private final float PI = 1.01f;
+	private final double KW = 1.01f;
+	private final double PD = 0.7f;
+	private final double PI = 1.01f;
 	
 	private static final int TO = 10;
-	public static final float DX = 15.0f/(TO-1);
+	public static final double DX = 15.0f/(TO-1);
 
 	protected Layer[] layers;
 
@@ -44,7 +44,7 @@ public class NeuralNet {
 	 *            1), +1 bo polaryzacja
 	 */
 
-	public NeuralNet(float learnRate, final int numOfLayers,
+	public NeuralNet(double learnRate, final int numOfLayers,
 			final int numNeurons1, final int numNeuron1Size,
 			final int numNeurons2) {
 		this.error = 0;
@@ -55,7 +55,7 @@ public class NeuralNet {
 		layers[1] = new Layer(numNeurons2, numNeurons1 + 1); // warstwa wyjsciowa (tyle ile w ukrytej + 1 (polaryzacja)
 	}
 
-	public void learnNet(float[] inputX, float target) {
+	public void learnNet(double[] inputX, double target) {
 		layers[0].startLayer(inputX, true);
 		layers[1].startLayer(layers[0].getLayerOutput(), false); // przekazuje wyjscie warstwy ukrytej
 																 // na wejscie warstwy wyjsciowej
@@ -69,23 +69,23 @@ public class NeuralNet {
 														// ZMIENIAM)
 	}
 	
-	public float goForward(float[] inputX) {
+	public double goForward(double[] inputX) {
 		layers[0].startLayer(inputX, true);
 		layers[1].startLayer(layers[0].getLayerOutput(), false);
 		return layers[1].getLastSolution();
 	}
 
-	public float goThroughLearning(float[] data) {
+	public double goThroughLearning(double[] data) {
 		// System.out.println(NN.toString());
 		/*
 		if (x.length != data.length) {
 			System.out.println("Rozmiar wektora wejsciowego != wektora z danymi");
 			return -1;
 		}*/
-		float[] x = new float[1];
+		double[] x = new double[1];
 		setErrorZero();
 		for (int i = 0; i < data.length; i++) {
-			x = new float[] { 0.1f+DX*i };
+			x = new double[] { 0.1f+DX*i };
 			//learnNet(x, data[i]);
 			// System.out.println(NN.toStringX());
 			setError(goForward(x), data[i]);
@@ -98,14 +98,14 @@ public class NeuralNet {
 		return getError();
 	}
 
-	public float goThroughLearning(Vector<float[]> dVec, float[] data) {
+	public double goThroughLearning(Vector<double[]> dVec, double[] data) {
 		// System.out.println(NN.toString());
 		/*
 		if (x.length != data.length) {
 			System.out.println("Rozmiar wektora wejsciowego != wektora z danymi");
 			return -1;
 		}*/
-		float[] x = new float[5];
+		double[] x = new double[5];
 		setErrorZero();
 		for (int i = 0; i < data.length; i++) {
 			x = dVec.get(i);
@@ -122,13 +122,13 @@ public class NeuralNet {
 	}
 	
 	/*
-	public float goForwardMinKierunkowa(float[] inputX, float alfa) {
+	public double goForwardMinKierunkowa(double[] inputX, double alfa) {
 		layers[0].startLayer(inputX, true);
 		layers[1].startLayer(layers[0].getLayerOutput(), false);
 	}
 	*/
 	
-	public void setError(float y, float d) { // blad sredniokwadratowy sieci
+	public void setError(double y, double d) { // blad sredniokwadratowy sieci
 		this.error = this.error + 0.5f*((y - d) * (y - d));
 	}
 
@@ -136,11 +136,11 @@ public class NeuralNet {
 		this.error = 0;
 	}
 
-	public float getError() {
+	public double getError() {
 		return error;
 	}
 
-	public float getLayerLastSolution() {
+	public double getLayerLastSolution() {
 		return layers[1].getLastSolution();
 	}
 
@@ -169,7 +169,7 @@ public class NeuralNet {
 		layers[0].updateWeightsInNeuronsLayer();
 	}
 	
-	public void updateWeightsInLayersGradSprzez(float paramGradSprzez) { // jako argument dac wart wsp uczenia
+	public void updateWeightsInLayersGradSprzez(double paramGradSprzez) { // jako argument dac wart wsp uczenia
 		// sie
 		// for (Layer l : layers) {
 		// l.updateWeightsInNeuronsLayer(this.learnRate);
@@ -183,7 +183,7 @@ public class NeuralNet {
 		layers[0].updateWeightsInNeuronsLayerGradSprzez(paramGradSprzez);
 	}
 	
-	public void updateCopyWeightsInLayers(float alfa) {
+	public void updateCopyWeightsInLayers(double alfa) {
 		for (Layer l : layers) {
 			l.updateCopyWeightsInNeuronsLayer(alfa);
 		}
@@ -220,24 +220,24 @@ public class NeuralNet {
 		this.prevError = this.error;
 	}
 
-	public float getLearnRate() {
+	public double getLearnRate() {
 		return this.learnRate;
 	}
 	
-	public void setLearnRate(float learnRate) {
+	public void setLearnRate(double learnRate) {
 		this.learnRate = learnRate;
 	}
 	
-	public List<Float> getWeightsChangeTable() {
-		List<Float> tmp = new ArrayList<Float>();
+	public List<Double> getWeightsChangeTable() {
+		List<Double> tmp = new ArrayList<Double>();
 		for (Layer l : layers) {
 			tmp.addAll(l.getWeightsChangeLayer());
 		}
 		return tmp;
 	}
 	
-	public List<Float> getPElementTable() {
-		List<Float> tmp = new ArrayList<Float>();
+	public List<Double> getPElementTable() {
+		List<Double> tmp = new ArrayList<Double>();
 		for (Layer l : layers) {
 			tmp.addAll(l.getPElementLayer());
 		}
@@ -247,7 +247,7 @@ public class NeuralNet {
 	/*-------------------------------------------------------------------------------*/
 	/*------------- USTALANIE WAG POCZATKOWYCH NA SZTYWNO DO TESTOW------------------*/
 	/* Ustalenie wag @newWeights w warstwie @indexL i neuronie @indexN---------------*/
-	public void setWeightsByParamInLayer(int indexL, int indexN, float[] newWeights) {
+	public void setWeightsByParamInLayer(int indexL, int indexN, double[] newWeights) {
 		layers[indexL].neurons[indexN].setWeightsByParam(newWeights);
 	}
